+++ tests/unit/sync/conftest.py
@@ -1,5 +1,8 @@
-from click.testing import CliRunner
 import logging
+
+from typer.testing import CliRunner
+
+from dbx.commands.sync.sync import sync_app
 
 
 def invoke_cli_runner(*args, **kwargs):
@@ -7,7 +10,7 @@ def invoke_cli_runner(*args, **kwargs):
     Helper method to invoke the CliRunner while asserting that the exit code is actually 0.
     """
     expected_error = kwargs.pop("expected_error") if "expected_error" in kwargs else None
-    res = CliRunner().invoke(*args, **kwargs)
+    res = CliRunner().invoke(sync_app, *args, **kwargs)
 
     if res.exit_code != 0:
         if not expected_error: