+++ tests/unit/sync/test_commands.py
@@ -4,7 +4,6 @@ from unittest.mock import patch, call, MagicMock
 import click
 import pytest
 
-from dbx.commands.sync.sync import dbfs, repo
 from dbx.commands.sync.functions import get_user_name, get_source_base_name
 from dbx.constants import DBX_SYNC_DEFAULT_IGNORES
 from dbx.sync import DeleteUnmatchedOption
@@ -34,7 +33,7 @@ def test_get_source_base_name():
 @patch("dbx.commands.sync.sync.main_loop")
 def test_repo_no_opts(mock_get_config, mock_main_loop):
     # some options are required
-    res = invoke_cli_runner(repo, [], expected_error=True)
+    res = invoke_cli_runner(["repo"], expected_error=True)
     assert "Missing option" in res.output
 
 
@@ -47,7 +46,7 @@ def test_repo_basic_opts(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_config.return_value = config
         mock_get_user_name.return_value = "me"
 
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -80,7 +79,7 @@ def test_repo_unknown_user(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_config.return_value = config
         mock_get_user_name.return_value = None
 
-        res = invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo"], expected_error=True)
+        res = invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo"], expected_error=True)
 
         assert mock_main_loop.call_count == 0
         assert mock_get_config.call_count == 1
@@ -96,7 +95,7 @@ def test_repo_dry_run(mock_get_config, mock_main_loop, mock_get_user_name):
     with temporary_directory() as tempdir:
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "--dry-run"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "--dry-run"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -129,7 +128,7 @@ def test_repo_polling(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_config.return_value = config
         mock_get_user_name.return_value = "me"
 
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "--polling-interval", "2"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "--polling-interval", "2"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -164,7 +163,7 @@ def test_repo_include_dir(mock_get_config, mock_main_loop, mock_get_user_name):
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -198,7 +197,7 @@ def test_repo_force_include_dir(mock_get_config, mock_main_loop, mock_get_user_n
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-fi", "foo"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-fi", "foo"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -232,7 +231,7 @@ def test_repo_include_pattern(mock_get_config, mock_main_loop, mock_get_user_nam
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-ip", "foo/*.py"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-ip", "foo/*.py"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -266,7 +265,7 @@ def test_repo_force_include_pattern(mock_get_config, mock_main_loop, mock_get_us
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-fip", "foo/*.py"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-fip", "foo/*.py"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -300,7 +299,7 @@ def test_repo_exclude_dir(mock_get_config, mock_main_loop, mock_get_user_name):
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-e", "foo"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-e", "foo"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -334,7 +333,7 @@ def test_repo_exclude_pattern(mock_get_config, mock_main_loop, mock_get_user_nam
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-ep", "foo/**/*.py"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-ep", "foo/**/*.py"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -370,7 +369,7 @@ def test_repo_include_dir_not_exists(mock_get_config, mock_main_loop, mock_get_u
 
         config = MagicMock()
         mock_get_config.return_value = config
-        res = invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo"], expected_error=True)
+        res = invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo"], expected_error=True)
 
         assert mock_main_loop.call_count == 0
         assert mock_get_config.call_count == 1
@@ -388,7 +387,7 @@ def test_repo_inferred_source(mock_get_config, mock_main_loop, mock_get_user_nam
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-d", "the-repo", "-u", "me"])
+        invoke_cli_runner(["repo", "-d", "the-repo", "-u", "me"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -422,7 +421,7 @@ def test_repo_inferred_source_no_git(mock_get_config, mock_main_loop, mock_get_u
 
         config = MagicMock()
         mock_get_config.return_value = config
-        res = invoke_cli_runner(repo, ["-d", "the-repo", "-u", "me"], expected_error=True)
+        res = invoke_cli_runner(["repo", "-d", "the-repo", "-u", "me"], expected_error=True)
 
         assert mock_main_loop.call_count == 0
         assert mock_get_config.call_count == 1
@@ -439,7 +438,9 @@ def test_repo_allow_delete_unmatched(mock_get_config, mock_main_loop, mock_get_u
         config = MagicMock()
         mock_get_config.return_value = config
 
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "--allow-delete-unmatched"])
+        invoke_cli_runner(
+            ["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "--unmatched-behaviour=allow-delete-unmatched"]
+        )
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -469,7 +470,9 @@ def test_repo_disallow_delete_unmatched(mock_get_config, mock_main_loop, mock_ge
         config = MagicMock()
         mock_get_config.return_value = config
 
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "--disallow-delete-unmatched"])
+        invoke_cli_runner(
+            ["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "--unmatched-behaviour=disallow-delete-unmatched"]
+        )
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -505,7 +508,11 @@ def test_dbfs_no_opts(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_user_name.return_value = "me"
 
         # we can run with no options as long as the source and user can be automatically inferred
-        invoke_cli_runner(dbfs, [])
+        invoke_cli_runner(
+            [
+                "dbfs",
+            ]
+        )
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -539,7 +546,7 @@ def test_dbfs_polling(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_user_name.return_value = "me"
 
         # we can run with no options as long as the source and user can be automatically inferred
-        invoke_cli_runner(dbfs, ["--polling-interval", "3"])
+        invoke_cli_runner(["dbfs", "--polling-interval", "3"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -574,7 +581,7 @@ def test_dbfs_dry_run(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_user_name.return_value = "me"
 
         # we can run with no options as long as the source and user can be automatically inferred
-        invoke_cli_runner(dbfs, ["--dry-run"])
+        invoke_cli_runner(["dbfs", "--dry-run"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -605,7 +612,7 @@ def test_dbfs_source_dest(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_user_name.return_value = "me"
 
         # we can run with no options as long as the source and user can be automatically inferred
-        invoke_cli_runner(dbfs, ["-s", tempdir, "-d", "/foo/bar"])
+        invoke_cli_runner(["dbfs", "-s", tempdir, "-d", "/foo/bar"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -638,7 +645,7 @@ def test_dbfs_specify_user(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_config.return_value = config
 
         # we can run with no options as long as the source and user can be automatically inferred
-        invoke_cli_runner(dbfs, ["-u", "someone"])
+        invoke_cli_runner(["dbfs", "-u", "someone"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -670,7 +677,7 @@ def test_dbfs_unknown_user(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_user_name.return_value = None
 
         # we can run with no options as long as the source and user can be automatically inferred
-        res = invoke_cli_runner(dbfs, [], expected_error=True)
+        res = invoke_cli_runner(["dbfs"], expected_error=True)
 
         assert mock_main_loop.call_count == 0
         assert mock_get_config.call_count == 1
@@ -689,7 +696,7 @@ def test_dbfs_no_root(mock_get_config, mock_main_loop, mock_get_user_name):
         mock_get_user_name.return_value = "me"
 
         # we can run with no options as long as the source and user can be automatically inferred
-        res = invoke_cli_runner(dbfs, ["-d", "/"], expected_error=True)
+        res = invoke_cli_runner(["dbfs", "-d", "/"], expected_error=True)
 
         assert mock_main_loop.call_count == 0
         assert mock_get_config.call_count == 1
@@ -712,7 +719,7 @@ def test_repo_use_gitignore(mock_get_config, mock_main_loop, mock_get_user_name)
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1
@@ -753,7 +760,7 @@ def test_repo_no_use_gitignore(mock_get_config, mock_main_loop, mock_get_user_na
 
         config = MagicMock()
         mock_get_config.return_value = config
-        invoke_cli_runner(repo, ["-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo", "--no-use-gitignore"])
+        invoke_cli_runner(["repo", "-s", tempdir, "-d", "the-repo", "-u", "me", "-i", "foo", "--no-use-gitignore"])
 
         assert mock_main_loop.call_count == 1
         assert mock_get_config.call_count == 1