+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED] - YYYY-MM-DD
 
-- Switch all of the CLI interfaces to `typer` (except `dbx sync`)
+- Switch all the CLI interfaces to `typer`
+- Arguments `--allow-delete-unmatched`/`--disallow-delete-unmatched` were **replaced** with `--unmatched-behaviour` option.
 
 ----
 > Unreleased changes must be tracked above this line.